backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
yfinance==0.2.36
pandas==2.2.3
python-dotenv==1.1.0
cors==1.0.1
alpha_vantage==3.0.0 